+++ client/client.go
@@ -4,6 +4,7 @@ import (
 	log "code.google.com/p/log4go"
 	"encoding/json"
 	"errors"
+	myrpc "github.com/Terry-Mao/gosnowflake/rpc"
 	"github.com/samuel/go-zookeeper/zk"
 	"math/rand"
 	"net/rpc"
@@ -20,8 +21,9 @@ const (
 	rpcClientPingSleep  = 1 * time.Second // rpc client ping need sleep
 	rpcClientRetrySleep = 1 * time.Second // rpc client retry connect need sleep
 
-	RPCPing   = "SnowflakeRPC.Ping"
-	RPCNextId = "SnowflakeRPC.NextId"
+	RPCPing    = "SnowflakeRPC.Ping"
+	RPCNextId  = "SnowflakeRPC.NextId"
+	RPCNextIds = "SnowflakeRPC.NextIds"
 )
 
 var (
@@ -100,6 +102,18 @@ func (c *Client) Id() (id int64, err error) {
 		return
 	}
 	if err = client.Call(RPCNextId, c.workerId, &id); err != nil {
+		log.Error("rpc.Call(\"%s\", %d, &id) error(%v)", RPCNextId, c.workerId, err)
+	}
+	return
+}
+
+// Ids generate a snowflake id.
+func (c *Client) Ids(num int) (ids []int64, err error) {
+	client, err := c.client()
+	if err != nil {
+		return
+	}
+	if err = client.Call(RPCNextIds, &myrpc.NextIdsArgs{WorkerId: c.workerId, Num: num}, &ids); err != nil {
 		log.Error("rpc.Call(\"%s\", %d, &id) error(%v)", RPCNextId, c.workerId, err)
 	}
 	return