+++ CHANGELOG.md
@@ -1,4 +1,14 @@
-## Version 1.2 (Pennding)
+## Version 1.3 (Pennding)
+## Version 1.2 
+
+New Features:
+    - Add NextIds for get mutiple ids.
+    - Add "start"(start timestamp) config for calculating snowflake id.
+
+Bugfixes:
+
+ - 
+
 ## Version 1.1 (2014-12-15)
 
 New Features: